requirements.txt
requests
beautifulsoup4
feedparser
google-genai
pydantic
SQLAlchemy
psycopg2-binary